Add a check cluster function (#39)
This change add a check cluster status function in the BeforeEach.

I also include a change on the stable sriov function if there is no
status I mark the cluster as not ready yet.

This change is because when a node came up after a reboot the config
daemon mark remove the status field but it doesn't mean the policy
is in synced status.

Signed-off-by: Sebastian Sch <sebassch@gmail.com>